Change axis order in CDF Estimator to store data contiguously in memory and thus improve performance
Benchmarking CDFEstimator makes clear, that storing the data in the last axis (`a[..., i]`) is not the way  to go
with numpy. This is caused by the internal order of numpy arrays being 'C-order' or row-major. Storing the data in the
first axis (`a[i, ...]`) instead cuts the computation time roughly in half
(from ~ 1.2 s to ~ 614 ms on my Macbook Pro, 13-inch, M1, 2020 for 100 1000x1000 images) and greatly facilitates the
indexing and thus increases the readability of the code.